Fixes #4 (changing from night to day takes a long time)
This is done using "busy loading", using to setup routines.
The first one is for drawing the splash screen.
Then with this one drawn in the screen already "busy loading" of the
sounds can be done without the user realizing and therefore not
degrading the user experience. Ideally though, this should be done with
a thread.