Sets new folder structure
This is in order to avoid having one repository for each platform and with a lot of duplicated and difficult to update code.

With this setup, the common code, the src/ folder and resources (images and audio), are shared for every platform. On top of that there is still room for specifying platform specific settings and files. In order to keep the src/ code linked between platforms they are symlinked and this way a change in Eclipse in the android project for example is automatically reflected in Xcode for the iOS project, all of that keeping the version control changes centralized.

There is also a linking.sh script for symlinking everything easier.