Fixes #27 (animation velocity too slow)
This also significantly improves the loading time of the app, since now
the files, instead of being all loaded to memory at the beginning of the
app, they are "streamed" while they are being played. Not sure if this
will work in all hardware though.

Also, important to add that for some very very weird reason the
streaming method of playing files seems to be working ok in the supposed same conditions as when before it
didn't. This is referring to my Moto X 2013, that didn't suffer any
software update and still using the exact same version (branch) of
openFrameworks. Guess could be somehow related with the conjuction of
calls of .pause(), .stop() and .loadSound() methods.